% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_location.R
\name{add_location}
\alias{add_location}
\title{add_location}
\usage{
add_location(x, locations, on_string = "serial")
}
\arguments{
\item{x}{data.table}

\item{locations}{data.table with file_name, in_water, and elevation}
}
\value{

}
\description{
add_location
}
