% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_location.R
\name{add_pore_pressure}
\alias{add_pore_pressure}
\title{to_pore_pressure}
\usage{
add_pore_pressure(x, dbar_to_m = NULL)
}
\arguments{
\item{x}{data.table}

\item{dbar_to_m}{value to use for dbar to m H2O conversion}
}
\description{
to_pore_pressure
}
