% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_start_end_times.R
\name{add_start_end_times}
\alias{add_start_end_times}
\title{add_start_end_times}
\usage{
add_start_end_times(x)
}
\arguments{
\item{x}{data.table of transducer data}
}
\value{
data.table with starting and end times for each dataset
}
\description{
add_start_end_times
}
