% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_time.R
\name{adjust_time}
\alias{adjust_time}
\title{adjust_time}
\usage{
adjust_time(x, n_sec = 0)
}
\arguments{
\item{x}{data.table of transducer data}

\item{n_sec}{number of seconds to adjust}
}
\value{

}
\description{
adjust_time
}
