% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_functions.R
\name{as_list}
\alias{as_list}
\title{as_list}
\usage{
as_list(acc, input)
}
\arguments{
\item{acc}{the accumulater, can be empty}

\item{input}{the input, can be empty}
}
\value{
A step function that yields a list
}
\description{
A normal reduction step function that accumulates values to a list.
It also supports arity 0 and 1.
}

