% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/step_functions.R
\name{as_vector}
\alias{as_vector}
\title{as_vector}
\usage{
as_vector(acc, input)
}
\arguments{
\item{acc}{the accumulater, can be empty}

\item{input}{the input, can be empty}
}
\value{
A step function that yields a vector
}
\description{
Accumulates values to a vector by appending results
}

