% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_manual.R
\name{compare_air}
\alias{compare_air}
\title{compare_air
This is for when all transducers should read the same value}
\usage{
compare_air(location, manual_wl, stat_fun = median)
}
\arguments{
\item{stat_fun}{}
}
\value{

}
\description{
compare_air
This is for when all transducers should read the same value
}
