% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_manual.R
\name{compare_manual}
\alias{compare_manual}
\title{compare_manual
This is for when all non-barometric transducers should read the same depth adjusted value}
\usage{
compare_manual(location, manual_wl, stat_fun = median)
}
\arguments{
\item{stat_fun}{}
}
\description{
compare_manual
This is for when all non-barometric transducers should read the same depth adjusted value
}
