% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\title{compose}
\usage{
compose(...)
}
\arguments{
\item{...}{functions that should be composed}
}
\value{
The composition of the functions.
}
\description{
Composes a number of functions. 
Evaluates functions from right to left.
}
\examples{
composition <- compose(
   map(function(x) x + 1),
   keep(function(x) x \%\% 3 == 0),
   keep(function(x) x < 500),
   take(5)
 )
}

