% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{\%.\%}
\alias{\%.\%}
\title{composition operator}
\usage{
a \%.\% b
}
\arguments{
\item{a}{a function that will be composed with function b.}

\item{b}{a function that takes the result of function b.}
}
\value{
The composition of the functions.
}
\description{
a %.% b translates into compose(a, b). 
It should be used to design pipes based on transducers.
}
\examples{
composition <- map(function(x) x + 1) \%.\%
   keep(function(x) x \%\% 3 == 0) \%.\%
   keep(function(x) x < 500) \%.\%
   take(5)
}

