% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/step_functions.R
\name{contains}
\alias{contains}
\title{contains}
\usage{
contains(f)
}
\arguments{
\item{f}{a function that takes an input and returns TRUE or FALSE.}
}
\value{
A transducer that checks for a value using f.
}
\description{
Checks if an element is in the sequence or not.
The process stopes when the element is contained in the sequence.
}

