% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequence.R
\name{create_sequence}
\alias{create_sequence}
\title{creates a lazy sequence}
\usage{
create_sequence(head_value, generator)
}
\arguments{
\item{head_value}{the value of the first element}

\item{generator}{a function that takes the current head and produces a new item}
}
\value{
A lazy sequence of items
}
\description{
Create an (infinite) sequence
}
\examples{
natural_numbers <- create_sequence(0, function(x) x + 1)
}

