% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequence.R
\name{create_sequence_from_function}
\alias{create_sequence_from_function}
\title{creates a lazy sequence from a arity-0 function}
\usage{
create_sequence_from_function(fn)
}
\arguments{
\item{fn}{a function that returns a value}
}
\value{
A lazy sequence created by calling fn whenver a new item is needed.
}
\description{
Create an (infinite) lazy sequence
}
\examples{
random_numbers <- create_sequence_from_function(function() runif(1))
}

