% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{density_temperature}
\alias{density_temperature}
\title{density_temperature}
\usage{
density_temperature(x)
}
\arguments{
\item{x}{temperature}
}
\description{
density_temperature
}
\details{
This function estimates the density of water at a given temperature
using Kell's formulation. ITS-90
}
\seealso{
\link{https://nvlpubs.nist.gov/nistpubs/jres/097/jresv97n3p335_A1b.pdf}
}
