% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transducers.R
\name{discard}
\alias{discard}
\title{discard}
\usage{
discard(f)
}
\arguments{
\item{f}{a function that takes an input and returns TRUE or FALSE.}
}
\value{
A transducer that discards a value using f.
}
\description{
Removes elements where the predicate is TRUE.
}

