% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transducers.R
\name{distinct}
\alias{distinct}
\title{distinct}
\usage{
distinct()
}
\value{
A transducer that filters out duplicate elements.
}
\description{
Removes duplicate elements. Run linear in terms of distinct elements in the sequence.
}

