% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{.check_files}
\alias{.check_files}
\title{.check_files}
\usage{
.check_files(x)
}
\arguments{
\item{x}{file path(s) of the transducers}
}
\value{
error for invalid inputs
}
\description{
.check_files
}
\details{
This function checks the input to be read to see if files exist
and the format is correct.
}
\examples{
.check_files(NULL)
.check_files(as.character(rnorm(1)))
.check_files(rnorm(1))
}
\keyword{internal}
