% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_times.R
\name{.check_times}
\alias{.check_times}
\title{.check_times}
\usage{
.check_times(x)
}
\arguments{
\item{x}{either a character string or datetime object}
}
\value{
error for invalid inputs
}
\description{
.check_times
}
\details{
This function checks the class of the datetime
}
\examples{
.check_times('2018-01-01')
}
\keyword{internal}
