% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_data.R
\name{downsample_data}
\alias{downsample_data}
\title{downsample_data}
\usage{
downsample_data(x, time_interval = 3600)
}
\arguments{
\item{x}{the data.table of transducer readings}

\item{time_interval}{the time interval in seconds}
}
\value{
data.table of subsampled data
}
\description{
downsample_data
}
