% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/step_functions.R
\name{filter}
\alias{filter}
\title{filter}
\usage{
filter(f)
}
\arguments{
\item{f}{a function that takes an input and returns TRUE or FALSE.}
}
\value{
A transducer that filters a value using f.
}
\description{
A filter transducer.
}

