% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gaps.R
\name{filter_duplicates}
\alias{filter_duplicates}
\title{filter_duplicates}
\usage{
filter_duplicates(x, fun = data.table::first, ...)
}
\arguments{
\item{x}{data.table with time and value columns}

\item{fun}{first, last, mean, median}

\item{...}{arguments to pass to fun}
}
\value{

}
\description{
filter_duplicates
}
\examples{
x <- data.table(datetime = sample(1:1000, 10000, replace = TRUE))
x[, value := rnorm(nrow(x))]
setkey(x, datetime)
system.time(filter_duplicates(x, mean))
}
