% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transducers.R
\name{flat_map}
\alias{flat_map}
\title{flat_map}
\usage{
flat_map(f)
}
\arguments{
\item{f}{a function that takes an input and returns a value}
}
\value{
A transducer that maps a value and flattens the result.
}
\description{
Maps each elements to a new domain and flattens the result
}

