% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sql.R
\name{generate_sql}
\alias{generate_sql}
\title{generate_sql}
\usage{
generate_sql(db, start = NULL, end = NULL, by = NULL)
}
\arguments{
\item{db}{database connection}
}
\value{
sql string

sql string
}
\description{
generate_sql
}
\examples{
generate_sql(Sys.time()-1000, Sys.time(), 10)


generate_sql

}
