% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_head.list}
\alias{get_head.list}
\title{Returns the first element of a list}
\usage{
\method{get_head}{list}(sequence)
}
\arguments{
\item{sequence}{a list}
}
\description{
Returns the first element of a list
}

