% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_head}
\alias{get_head}
\title{Returns the first element of a collection.}
\usage{
get_head(sequence)
}
\arguments{
\item{sequence}{a sequence}
}
\description{
Returns the first element of a collection.
}

