% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_head.transducer_sequence}
\alias{get_head.transducer_sequence}
\title{Returns the first element of a lazy sequence.}
\usage{
\method{get_head}{transducer_sequence}(sequence)
}
\arguments{
\item{sequence}{a sequence}
}
\description{
Returns the first element of a lazy sequence.
}

