% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity.R
\name{gravity}
\alias{gravity}
\title{gravity}
\usage{
gravity(elevation = 0, latitude = 0)
}
\arguments{
\item{elevation}{the elevation of the monitoring location in meters}

\item{latitude}{the latitude for the monitoring location in degrees}
}
\value{

}
\description{
gravity
}
\examples{
gravity()
gravity(1000, 45)
}
\seealso{
\link{https://www.sensorsone.com/local-gravity-calculator/}
}
