% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{hydrograph.data.table}
\alias{hydrograph.data.table}
\title{hydrograph

create a hydrograph}
\usage{
\method{hydrograph}{data.table}(x, var_type = "pressure", plot_col = "value", max_n = 10000)
}
\arguments{
\item{plot}{}
}
\description{

}
\details{

}
