% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transducers.R
\name{map}
\alias{map}
\title{map}
\usage{
map(f)
}
\arguments{
\item{f}{a function that takes an input and returns a value}
}
\value{
A transducer that maps a value using f.
}
\description{
A map transducer.
}

