% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transducers.R
\name{map_indexed}
\alias{map_indexed}
\title{map_indexed}
\usage{
map_indexed(f)
}
\arguments{
\item{f}{a function that takes an index and an input and returns a value}
}
\value{
A transducer that maps a value using f.
}
\description{
A map transducer that keeps an index. It starts with 0.
}

