% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micron.R
\name{micron_calibration}
\alias{micron_calibration}
\title{micron_calibration}
\usage{
micron_calibration(
  x,
  conversion_constant = 0.70307,
  units = "meters",
  robust = TRUE
)
}
\arguments{
\item{x}{file path}

\item{conversion_constant}{conversion constant default is psi to m H2O}

\item{units}{after conversion}

\item{robust}{use MASS::rlm method instead of lm}
}
\value{

}
\description{
micron_calibration
}
