% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequence.R
\name{next_value.list}
\alias{next_value.list}
\title{returns the next value of a list}
\usage{
\method{next_value}{list}(sequence)
}
\arguments{
\item{sequence}{a collection}
}
\description{
I.e. the second value.
}

