% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{next_value}
\alias{next_value}
\title{Takes the next value of collection}
\usage{
next_value(sequence)
}
\arguments{
\item{sequence}{a collection}
}
\value{
the next element in the collection
}
\description{
Takes the next value of collection
}
\examples{
second_random_number <- next_value(create_sequence_from_function(function() runif(1)))
}

