% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{next_value.transducer_sequence}
\alias{next_value.transducer_sequence}
\title{returns the next value of a lazy sequence}
\usage{
\method{next_value}{transducer_sequence}(sequence)
}
\arguments{
\item{sequence}{a sequence}
}
\description{
returns the next value of a lazy sequence
}

