% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_files.R
\name{parse_paths}
\alias{parse_paths}
\title{parse_paths}
\usage{
parse_paths(folder_path, ...)
}
\arguments{
\item{folder_path}{the folder to export files to}
}
\value{
list of file names
}
\description{
Get a vector of file names from a path
}
\examples{
pp <- c('/media/kennel/Seagate Expansion Drive/rbr_g360/', '/media/kennel/Seagate Expansion Drive/rbr_ssfl/077611_20140827_1025.rsk')
pp <- parse_paths(pp, full.names = TRUE, pattern = '.rsk')
}
