% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/step_functions.R
\name{plus}
\alias{plus}
\title{plus}
\usage{
plus(acc, input)
}
\arguments{
\item{acc}{the accumulater, can be empty}

\item{input}{the input, can be empty}
}
\value{
A step function that yields a sum
}
\description{
A normal reduction step function that sums over all items.
It also supports arity 0 and 1.
}

