% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transducers.R
\name{random_sample}
\alias{random_sample}
\title{random_sample}
\usage{
random_sample(prob, seed = NULL)
}
\arguments{
\item{prob}{a probability that an element is kept}

\item{seed}{a seed, optional.}
}
\value{
A transducer that keeps elements according to a probability.
}
\description{
Samples from the collection according to a probability.
}

