% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbr_undo_calibration.R
\name{rbr_estimate_millivolt}
\alias{rbr_estimate_millivolt}
\title{Remove the calibration parameters from the rbr to get raw readings}
\usage{
rbr_estimate_millivolt(rbr)
}
\arguments{
\item{rbr}{}
}
\description{
Remove the calibration parameters from the rbr to get raw readings
}
