% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbr_info.R
\name{rbr_info.character}
\alias{rbr_info.character}
\title{rbr_info}
\usage{
\method{rbr_info}{character}(x)
}
\arguments{
\item{x}{path(s) to .rsk file}
}
\value{
list of results
}
\description{
get supplementary info from .rsk file
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
