% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbr_start_end.R
\name{rbr_start_end.data.table}
\alias{rbr_start_end.data.table}
\title{rbr_start_end
start and end times for rsk data table}
\usage{
\method{rbr_start_end}{data.table}(x)
}
\arguments{
\item{x}{}
}
\description{
rbr_start_end
start and end times for rsk data table
}
