% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diver.R
\name{read_diver.ascii}
\alias{read_diver.ascii}
\title{read_diver}
\usage{
\method{read_diver}{ascii}(x, transducer_depth = NULL, well_elevation = NULL, ...)
}
\arguments{
\item{head_size}{}
}
\value{
data.table of results
}
\description{
read_diver
}
