% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diver.R
\name{read_diver.binary}
\alias{read_diver.binary}
\title{read_diver}
\usage{
\method{read_diver}{binary}(x, head_size = 280, ...)
}
\arguments{
\item{head_size}{}
}
\value{
data.table of results
}
\description{
read_diver
}
