% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diver.R
\name{read_diver.character}
\alias{read_diver.character}
\title{read_diver}
\usage{
\method{read_diver}{character}(x, ...)
}
\arguments{
\item{db_name}{character the path to the rbr database ( rsk )}
}
\value{
data.table of results
}
\description{
read_diver
}
