% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diver.R
\name{read_diver}
\alias{read_diver}
\title{obtain data from diver .mon file}
\usage{
read_diver(x, ...)
}
\arguments{
\item{x}{character the path to the rbr database ( rsk )}

\item{...}{}
}
\description{
import diver data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
