% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_levelogger.R
\name{read_levelogger}
\alias{read_levelogger}
\title{read_levelogger}
\usage{
read_levelogger(x, transducer_depth = NULL, well_elevation = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
read_levelogger
}
