% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micron.R
\name{read_micron}
\alias{read_micron}
\title{read_micron}
\usage{
read_micron(
  x,
  transducer_depth = NULL,
  well_elevation = NULL,
  calibration = NULL,
  ...
)
}
\arguments{
\item{x}{the full path to the file}

\item{calibration}{data table with calibration points}

\item{...}{}
}
\value{

}
\description{
read_micron
}
