% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rbr.R
\name{read_rbr.character}
\alias{read_rbr.character}
\title{read_rbr}
\usage{
\method{read_rbr}{character}(
  x,
  start = NULL,
  end = NULL,
  by = NULL,
  times = NULL,
  transducer_depth = NULL,
  well_elevation = NULL,
  ...
)
}
\arguments{
\item{start}{starting date to get}

\item{end}{ending date to get}

\item{by}{spacing}

\item{db_name}{character the path to the rbr database ( rsk )}

\item{use_rbr_tz}{boolean use time zone information from the rbr file?}
}
\value{
data.table of results
}
\description{
read_rbr
}
\examples{
db_name <- system.file("extdata", "example.rsk", package="transducer")
system.time(wl <- read_rbr(db_name))
system.time(wl <- read_rbr(db_name, by = 3600))
system.time(wl <- read_rbr(db_name, by = 1800))
system.time(wl <- read_rbr(db_name, by = 900))
system.time(wl <- read_rbr(db_name, by = 600))
system.time(wl <- read_rbr(db_name, by = 300))
system.time(wl <- read_rbr(db_name, by = 10))

db_name <- '/media/kennel/Data/phd/personnel/pat/syngenta/SYN UW1 P1-07 - 077924_20150413_1746.rsk'
system.time(wl <- read_rbr(db_name))
system.time(wl <- read_rbr(db_name,
 start = as.POSIXct('2015-01-01', tz = 'UTC'),
 end = as.POSIXct('2015-02-01', tz = 'UTC'),
 by = 60))

db_name <- '/media/kennel/Data/tmp/rd10 077611_20181108_1037.rsk'
system.time(wl <- read_rbr(db_name))

}
