% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rbr.R
\name{read_rbr.data.table}
\alias{read_rbr.data.table}
\title{read_rbr}
\usage{
\method{read_rbr}{data.table}(locations, start = NULL, end = NULL,
  by = NULL, times = NULL, ...)
}
\arguments{
\item{locations}{file with port information}

\item{start}{starting date to get}

\item{end}{ending date to get}

\item{by}{spacing}

\item{use_rbr_tz}{boolean use time zone information from the rbr file?}
}
\value{
data.table of results
}
\description{
read_rbr
}
