% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rbr_db.R
\name{read_rbr_db}
\alias{read_rbr_db}
\title{obtain data from rbr sqlite3 database}
\usage{
read_rbr_db(db, db_name, sql_text)
}
\arguments{
\item{db}{database connection}

\item{sql_text}{character sql string to execute on connection for filtering}
}
\value{
data.table of results
}
\description{
import sqlite data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
