% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rbr_raw.R
\name{read_rbr_raw}
\alias{read_rbr_raw}
\title{read_rbr_raw}
\usage{
read_rbr_raw(x, type = c("raw", "basic", "compensated"))
}
\arguments{
\item{x}{(character) name of the rsk file}

\item{type}{(character vector) one or more of 'raw', 'basic', or 'compensated'}
}
\value{

}
\description{
read_rbr_raw
}
