% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{rest}
\alias{rest}
\title{Returns the rest of a collection}
\usage{
rest(sequence)
}
\arguments{
\item{sequence}{a collection

That is everything except for the first item. 
Returns NULL if no more items exist.}
}
\description{
Returns the rest of a collection
}

