% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequence.R
\name{rest.transducer_sequence}
\alias{rest.transducer_sequence}
\title{Returns the rest of a lazy sequence.}
\usage{
\method{rest}{transducer_sequence}(sequence)
}
\description{
Returns the rest of a lazy sequence.
}

