% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_data.R
\name{subset_time_interval}
\alias{subset_time_interval}
\title{subset_time_interval}
\usage{
subset_time_interval(x, start = NULL, end = NULL)
}
\arguments{
\item{x}{the data.table of transducer readings}

\item{start}{beginning of subset (POSIXct)}

\item{end}{end of subset (POSIXct)}
}
\value{
data.table with values between start and end
}
\description{
subset_time_interval
}
