% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_files.R
\name{summarize_files}
\alias{summarize_files}
\title{summarize_files}
\usage{
summarize_files(file_names, output_folder = "")
}
\arguments{
\item{file_names}{the folder to read}

\item{output_folder}{the folder to export files to}
}
\description{
This function reads in a folder of rbr files and generates a summary
}
\examples{
folder_path <- '/media/kennel/Seagate Expansion Drive/rbr_ssfl'
tmp <- summarize_files(folder_path)
}
