% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transducers.R
\name{take}
\alias{take}
\title{take}
\usage{
take(number_of_results)
}
\arguments{
\item{number_of_results}{a number indicating how many items this function should accept.}
}
\value{
A transducer that stops after number_of_results items.
}
\description{
Takes x number of items from a process.
}

