% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transducers.R
\name{take_nth}
\alias{take_nth}
\title{take}
\usage{
take_nth(n)
}
\arguments{
\item{n}{indicates the what nth element.}
}
\value{
A transducer that takes every nth element
}
\description{
Takes every nth element.
}

