% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/step_functions.R
\name{take_while}
\alias{take_while}
\title{take_while}
\usage{
take_while(f)
}
\arguments{
\item{f}{a predicate indicating how long items should be processed.}
}
\value{
A transducer that stops after f evaluates to FALSE.
}
\description{
Take an item as long as a predicate is TRUE.
}

