% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compose.R
\name{\%then\%}
\alias{\%then\%}
\title{composition operator}
\usage{
a \%then\% b
}
\arguments{
\item{a}{a function that will be composed with function b.}

\item{b}{a function that takes the result of function b.}
}
\value{
The composition of the functions.
}
\description{
a %then% b translates into compose(a, b).
It should be used to design pipes based on transducers.
}
\examples{
composition <- map(function(x) x + 1) \%then\%
   keep(function(x) x \%\% 3 == 0) \%then\%
   keep(function(x) x < 500) \%then\%
   take(5)
}

