% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{to_list.transducer_sequence}
\alias{to_list.transducer_sequence}
\title{Converts a lazy sequence to a list}
\usage{
\method{to_list}{transducer_sequence}(sequence)
}
\arguments{
\item{sequence}{a list}
}
\description{
Converts a lazy sequence to a list
}

