% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transduce.R
\name{transduce}
\alias{transduce}
\title{Transduce}
\usage{
transduce(transducer, step, collection, init = NULL)
}
\arguments{
\item{transducer}{a transducer.}

\item{step}{an step function.}
}
\value{
The result of the reduction piped through the transducer.
}
\description{
This function is like reduce but it takes a transducer and a normal
step function used in reduce.
}
\details{
More documentation will follow. Api not stable yet.
}

