% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transduce.R
\name{transduce_tbl}
\alias{transduce_tbl}
\title{Transduce_tbl}
\usage{
transduce_tbl(transducer, step, tbl, margin = "rows",
  init = transducer(step)())
}
\arguments{
\item{transducer}{a transducer.}

\item{step}{an step function.}

\item{tbl}{a table.}

\item{margin}{indicates wether to iterate over "rows" or "columns". Default value is "rows".}

\item{init}{an initial value. If ommitted the step function will be called to generate a value without an argument.}
}
\value{
The result of the reduction piped through the transducer.
}
\description{
Similiar to the normal transduce function but it explicitly only works for
data.frames or matrices. It has an additional parameter that indicates if
it should be iterated over columns or rows.
By default it iterates over the rows starting with the first row.
}

