% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transducer.R
\docType{package}
\name{transducer}
\alias{transducer}
\alias{transducer-package}
\title{transducer}
\description{
Implements transducers introduced in Clojure as well as infinite sequences.
}
\examples{
transduce(take(5), plus, create_sequence(0, function(x) x + 1), init = 0)
}

