% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_data.R
\name{unnest_data}
\alias{unnest_data}
\title{unnest_data
This function pulls the data and auxilary columns out of a nested data.table}
\usage{
unnest_data(x, columns = c("serial", "type", "is_baro"))
}
\arguments{
\item{x}{the nested data.table of values}

\item{columns}{the column names to include with the data}
}
\description{
unnest_data
This function pulls the data and auxilary columns out of a nested data.table
}
