% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{vertigraph.data.table}
\alias{vertigraph.data.table}
\title{vertigraph

Creates a vertical profile}
\usage{
\method{vertigraph}{data.table}(x, var_type = "pressure", plot_col = "value", max_n = 10000)
}
\arguments{
\item{plot}{}
}
\value{

}
\description{

}
\details{

}
