% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westbay_to_elevation.R
\name{westbay_make_regular}
\alias{westbay_make_regular}
\title{westbay_make_regular}
\usage{
westbay_make_regular(x, max_difference = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
westbay_make_regular
}
