% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westbay_to_elevation.R
\name{westbay_to_elevation.data.table}
\alias{westbay_to_elevation.data.table}
\title{westbay_to_elevation}
\usage{
\method{westbay_to_elevation}{data.table}(x, baro_port = 0)
}
\arguments{
\item{baro_port}{}
}
\description{
westbay_to_elevation
}
