% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westbay_to_elevation.R
\name{westbay_to_elevation.numeric}
\alias{westbay_to_elevation.numeric}
\title{westbay_to_elevation}
\usage{
\method{westbay_to_elevation}{numeric}(x, transducer_depth, port_depth, data, units, ...)
}
\arguments{
\item{...}{}
}
\description{
westbay_to_elevation
}
