% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rbr.R
\name{write_rbr}
\alias{write_rbr}
\title{write_rbr}
\usage{
write_rbr(x, folder_path, datetime_format = "write.csv", ...)
}
\arguments{
\item{x}{output from read_rbr}

\item{folder_path}{the folder to export files to}

\item{datetime_format}{one of "ISO","squash","epoch","write.csv"}

\item{...}{arguments to pass to fwrite}
}
\value{

}
\description{
write_rbr
}
