% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rbr.R
\name{write_rbr_info}
\alias{write_rbr_info}
\title{write_rbr_info}
\usage{
write_rbr_info(x, file_path, ...)
}
\arguments{
\item{x}{the data.table of data}

\item{file_path}{the output file path}

\item{...}{}
}
\description{
write_rbr_info
}
