% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_location.R
\name{add_water_level}
\alias{add_water_level}
\title{add_water_level}
\usage{
add_water_level(x, baro_serial = NULL, dbar_to_m = 1.019716, ...)
}
\arguments{
\item{...}{}
}
\description{
If baro_serial is not provided this function will try to use a baro logger in
the provided dataset.
}
