% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbr_info.R
\name{rbr_info}
\alias{rbr_info}
\title{rbr_info}
\usage{
rbr_info(x, ...)
}
\arguments{
\item{x}{path to .rsk file}

\item{db}{database connection}

\item{db_name}{character the path to the rbr database ( rsk )}
}
\value{
list of results
}
\description{
get supplementary info from .rsk file
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
